+++ Tests/SystemTests.cs
@@ -50,6 +50,34 @@ namespace Tests
         }
 
         [Fact]
+        public async Task StreamSystem_SupportsAwaitUsing()
+        {
+            var config = new StreamSystemConfig
+            {
+                Endpoints = new List<EndPoint>
+                {
+                    new IPEndPoint(IPAddress.Loopback, 9999), // this should not be bound
+                    new IPEndPoint(IPAddress.Loopback, 5552),
+                },
+                ClientProvidedName = "StreamSystem_SupportsAwaitUsing"
+            };
+            await using (var system = await StreamSystem.Create(config))
+            {
+                await SystemUtils.WaitUntilAsync(() => SystemUtils.IsConnectionOpen("StreamSystem_SupportsAwaitUsing"));
+                Assert.False(system.IsClosed);
+            }
+
+            await SystemUtils.WaitUntilAsync(() => (SystemUtils.ConnectionsCountByName("StreamSystem_SupportsAwaitUsing").Result == 0));
+
+            // After await using block, DisposeAsync was called; create another system to confirm no resource leak
+            await using (var system2 = await StreamSystem.Create(config))
+            {
+                await SystemUtils.WaitUntilAsync(() => SystemUtils.IsConnectionOpen("StreamSystem_SupportsAwaitUsing"));
+                Assert.False(system2.IsClosed);
+            }
+        }
+
+        [Fact]
         public async Task CreateSystemThrowsWhenNoEndpointsAreReachable()
         {
             var config = new StreamSystemConfig
@@ -59,8 +87,10 @@ namespace Tests
                     new IPEndPoint(IPAddress.Loopback, 9999) // this should not be bound
                 }
             };
-            await Assert.ThrowsAsync<StreamSystemInitialisationException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<StreamSystemInitialisationException>(async () =>
+                {
+                    await StreamSystem.Create(config);
+                }
             );
         }
 
@@ -78,8 +108,7 @@ namespace Tests
                                              SslPolicyErrors.RemoteCertificateNameMismatch
                 },
             };
-            await Assert.ThrowsAsync<SslException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<SslException>(async () => { await StreamSystem.Create(config); }
             );
         }
 
@@ -142,8 +171,10 @@ namespace Tests
         public async Task CreateSystemThrowsWhenVirtualHostFailureAccess()
         {
             var config = new StreamSystemConfig { VirtualHost = "DOES_NOT_EXIST" };
-            await Assert.ThrowsAsync<VirtualHostAccessFailureException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<VirtualHostAccessFailureException>(async () =>
+                {
+                    await StreamSystem.Create(config);
+                }
             );
         }
 
@@ -151,8 +182,7 @@ namespace Tests
         public async Task CreateSystemThrowsWhenAuthenticationAccess()
         {
             var config = new StreamSystemConfig { UserName = "user_does_not_exist" };
-            await Assert.ThrowsAsync<AuthenticationFailureException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<AuthenticationFailureException>(async () => { await StreamSystem.Create(config); }
             );
         }
 
@@ -172,8 +202,10 @@ namespace Tests
             var config = new StreamSystemConfig { UserName = "guest", Password = "guest" }; // specified for readability
             var streamSystem = await StreamSystem.Create(config);
 
-            await Assert.ThrowsAsync<AuthenticationFailureException>(
-                async () => { await streamSystem.UpdateSecret("not_valid_secret"); }
+            await Assert.ThrowsAsync<AuthenticationFailureException>(async () =>
+                {
+                    await streamSystem.UpdateSecret("not_valid_secret");
+                }
             );
             await streamSystem.Close();
         }
@@ -206,8 +238,10 @@ namespace Tests
             var system = await StreamSystem.Create(config);
             await system.CreateStream(new StreamSpec(stream) { MaxLengthBytes = 20, });
 
-            await Assert.ThrowsAsync<CreateStreamException>(
-                async () => { await system.CreateStream(new StreamSpec(stream) { MaxLengthBytes = 10000, }); }
+            await Assert.ThrowsAsync<CreateStreamException>(async () =>
+                {
+                    await system.CreateStream(new StreamSpec(stream) { MaxLengthBytes = 10000, });
+                }
             );
             await system.DeleteStream(stream);
             await system.Close();
@@ -223,20 +257,28 @@ namespace Tests
             var config = new StreamSystemConfig();
             var system = await StreamSystem.Create(config);
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QueryOffset(string.Empty, "stream_we_don_t_care"); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QueryOffset(string.Empty, "stream_we_don_t_care");
+                }
             );
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QueryOffset("reference_we_don_care", string.Empty); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QueryOffset("reference_we_don_care", string.Empty);
+                }
             );
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QueryOffset(string.Empty, string.Empty); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QueryOffset(string.Empty, string.Empty);
+                }
             );
 
-            await Assert.ThrowsAsync<QueryException>(
-                async () => { await system.QueryPartition("stream_does_not_exist"); }
+            await Assert.ThrowsAsync<QueryException>(async () =>
+                {
+                    await system.QueryPartition("stream_does_not_exist");
+                }
             );
 
             await system.Close();
@@ -251,16 +293,22 @@ namespace Tests
             var config = new StreamSystemConfig();
             var system = await StreamSystem.Create(config);
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QuerySequence(string.Empty, "stream_we_don_t_care"); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QuerySequence(string.Empty, "stream_we_don_t_care");
+                }
             );
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QuerySequence("reference_we_don_care", string.Empty); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QuerySequence("reference_we_don_care", string.Empty);
+                }
             );
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.QuerySequence(string.Empty, string.Empty); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.QuerySequence(string.Empty, string.Empty);
+                }
             );
             await system.Close();
         }
@@ -271,8 +319,10 @@ namespace Tests
             // the user can set the SALs configuration externally
             // this test validates that the configuration is supported by the server
             var config = new StreamSystemConfig() { AuthMechanism = AuthMechanism.External };
-            await Assert.ThrowsAsync<AuthMechanismNotSupportedException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<AuthMechanismNotSupportedException>(async () =>
+                {
+                    await StreamSystem.Create(config);
+                }
             );
         }
 
@@ -280,8 +330,7 @@ namespace Tests
         public async Task ValidateRpCtimeOut()
         {
             var config = new StreamSystemConfig() { RpcTimeOut = TimeSpan.FromMilliseconds(1) };
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await StreamSystem.Create(config); }
+            await Assert.ThrowsAsync<ArgumentException>(async () => { await StreamSystem.Create(config); }
             );
         }
 
@@ -383,24 +432,50 @@ namespace Tests
             var system = await StreamSystem.Create(new StreamSystemConfig());
             var specZeroPartitions = new PartitionsSuperStreamSpec(SuperStream, 0) { };
 
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.CreateSuperStream(specZeroPartitions); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.CreateSuperStream(specZeroPartitions);
+                }
             );
 
             var specNullBindings = new BindingsSuperStreamSpec(SuperStream, null);
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.CreateSuperStream(specNullBindings); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.CreateSuperStream(specNullBindings);
+                }
             );
 
             var specDuplicationBindings =
                 new BindingsSuperStreamSpec(SuperStream, new[] { "duplication", "duplication" });
-            await Assert.ThrowsAsync<ArgumentException>(
-                async () => { await system.CreateSuperStream(specDuplicationBindings); }
+            await Assert.ThrowsAsync<ArgumentException>(async () =>
+                {
+                    await system.CreateSuperStream(specDuplicationBindings);
+                }
             );
 
-            await Assert.ThrowsAsync<DeleteStreamException>(
-                async () => { await system.DeleteSuperStream("not-exist"); }
+            await Assert.ThrowsAsync<DeleteStreamException>(async () => { await system.DeleteSuperStream("not-exist"); }
             );
+            await system.Close();
+        }
+
+        [Fact]
+        public async Task ServerPropertiesShouldBeNotEmpty()
+        {
+            var system = await StreamSystem.Create(new StreamSystemConfig());
+            var properties = await system.ServerProperties();
+            Assert.NotEmpty(properties);
+            Assert.Contains("product", properties);
+            Assert.Contains("platform", properties);
+            Assert.Contains("cluster_name", properties);
+            Assert.Contains("information", properties);
+            Assert.Contains("copyright", properties);
+            Assert.Contains("RabbitMQ", properties["product"]);
+            Assert.Contains("rabbit", properties["cluster_name"]);
+            Assert.Contains("Erlang", properties["platform"]);
+            Assert.Contains("rabbitmq.com", properties["information"]);
+            Assert.Contains("Copyright", properties["copyright"]);
+            Assert.Contains("version", properties);
+
             await system.Close();
         }
 
@@ -415,6 +490,54 @@ namespace Tests
             Assert.Equal((ulong)4, await system.QueryOffset(consumerRef, stream));
             await system.DeleteStream(stream);
             await system.Close();
+        }
+
+        [Fact]
+        public async Task CreateSystemWithCustomLookupLocatorStrategy()
+        {
+            var expectedDelay = TimeSpan.FromMilliseconds(500);
+            var customStrategy = new TestLookupLocatorStrategy(5, expectedDelay);
+            var config = new StreamSystemConfig
+            {
+                Endpoints = new List<EndPoint>
+                {
+                    new IPEndPoint(IPAddress.Loopback, 9999),
+                    new IPEndPoint(IPAddress.Loopback, 5552),
+                },
+                LookupLocatorStrategy = customStrategy
+            };
+            var system = await StreamSystem.Create(config);
+            Assert.False(system.IsClosed);
+            Assert.Equal(5, customStrategy.MaxAttempts);
+            Assert.Equal(expectedDelay, customStrategy.Delay);
+            await system.Close();
+            Assert.True(system.IsClosed);
+        }
+
+        private sealed class TestLookupLocatorStrategy : ILookupLocatorStrategy
+        {
+            public TestLookupLocatorStrategy(int maxAttempts, TimeSpan delay)
+            {
+                MaxAttempts = maxAttempts;
+                Delay = delay;
+            }
+
+            public int MaxAttempts { get; init; }
+            public TimeSpan Delay { get; }
+        }
+
+        [Fact]
+        public async Task UseDnsEndpointShouldWork()
+        {
+            var config = new StreamSystemConfig()
+            {
+                Endpoints = new List<EndPoint> { new DnsEndPoint("localhost", 5552) }
+            };
+            var system = await StreamSystem.Create(config);
+            Assert.False(system.IsClosed);
+            await system.Close();
+            Assert.True(system.IsClosed);
+
         }
     }
 }