+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.11.2]
+### 1.11.2 - 2026-18-05
+- [Release notes 1.11.2](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/releases/tag/v1.11.2)
+### Enhancements
+- Expose server ServerProperties by @Gsantomaggio in [#465](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/465)
+
+## [1.11.0]
+### 1.11.0 - 2026-17-03
+- [Release notes 1.11.0](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/releases/tag/v1.11.0)
+
+### Enhancements
+- Improve connection pool by @Gsantomaggio in [#449](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/449)
+- feat: implement IAsyncDisposable on StreamSystem by @Gsantomaggio in [#457](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/457)
+- The ConfirmationPipe improvement by @Gsantomaggio in [#458](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/458)
+- Add StatusChanged docs, locator reconnection retries by @Gsantomaggio in [#459](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/459)
+- Improve client documentation by @Gsantomaggio in [#460](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/460)
+- Add client documentation by @Gsantomaggio in [#461](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/461)
+
+### Bug Fix
+- fix stream filter example dependencies updated by @Gsantomaggio in [#452](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/pull/452)
+
 ## [1.10.0]
 ### 1.10.0 - 2026-14-01
 - [Release notes 1.10.0](https://github.com/rabbitmq/rabbitmq-stream-dotnet-client/releases/tag/v1.10.0)